Refactor : 시그널 함수 TimelineManager에서 호출할 수 있도록 변경
