+++ Assets/Timeline/Signal/LerpTimeScale.signal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9394698ad6c63b64e9c42a1b19f54b88
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
